requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
pydantic==1.10.7
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
gspread==5.9.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
openpyxl==3.1.2
fpdf2==2.7.4
python-dotenv==1.0.0
aiofiles==23.1.0
httpx==0.24.0
supabase==1.0.3
gotrue==1.0.1
realtime==1.0.0
starlette==0.26.1
alembic==1.10.4